+++ VkStorageBuffer/VulkanMeshRenderer.cpp
@@ -410,7 +410,7 @@ void FVulkanMeshRenderer::CreateStorageBuffers()
 		std::vector<FStorageBufferInfo>& StorageBuffers = Pair.second.StorageBuffers;
 		std::vector<FVulkanModel*>& Models = Pair.second.Models;
 
-		uint32_t StorageBufferSize = sizeof(FStorageBufferData) * Models.size();
+		uint32_t StorageBufferSize = static_cast<uint32_t>(sizeof(FStorageBufferData) * Models.size());
 
 		StorageBuffers.resize(MAX_CONCURRENT_FRAME);
 		for (int Idx = 0; Idx < MAX_CONCURRENT_FRAME; ++Idx)
@@ -673,6 +673,6 @@ void FVulkanMeshRenderer::Render()
 
 		vkCmdBindIndexBuffer(CommandBuffer, Mesh->GetIndexBuffer().Buffer, 0, VK_INDEX_TYPE_UINT32);
 
-		vkCmdDrawIndexed(CommandBuffer, static_cast<uint32_t>(MeshAsset->GetIndices().size()), Models.size(), 0, 0, 0);
+		vkCmdDrawIndexed(CommandBuffer, static_cast<uint32_t>(MeshAsset->GetIndices().size()), static_cast<uint32_t>(Models.size()), 0, 0, 0);
 	}
 }
